Less intrusive I/O operation readiness signaling
Instead of piggy-backing on a semi-standardized symbol that might not
be available in all libc implementations, use a custom function to
signal an I/O operation readiness by a file descriptor backend and
potentially wake up a thread blocked on a select(), pselect(), poll()
or ppoll() call.

The presence of the custom function is checked at link time (and the
mechanism is gracefully ignored if not present) in order to avoid tight
coupling between lwIP and the client implementation of the select(),
pselect(), poll() and ppoll() functionality.

Change-Id: I2ebc6fa77d725d6dc9f12df65d6a2cbf9468a073